# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/suppliers": {
      "post": {
        "operationId": "SuppliersController_create",
        "summary": "Создать поставщика",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Поставщик успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          }
        },
        "tags": [
          "suppliers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "SuppliersController_findAll",
        "summary": "Получить список поставщиков",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Количество элементов на странице",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Смещение для пагинации",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Поисковый запрос",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Направление сортировки: ASC или DESC",
            "schema": {
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список поставщиков получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedSupplierResponse"
                }
              }
            }
          }
        },
        "tags": [
          "suppliers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/suppliers/{id}": {
      "get": {
        "operationId": "SuppliersController_findOne",
        "summary": "Получить поставщика по ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Поставщик найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "404": {
            "description": "Поставщик не найден"
          }
        },
        "tags": [
          "suppliers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "SuppliersController_update",
        "summary": "Обновить поставщика",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupplierDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Поставщик успешно обновлен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Supplier"
                }
              }
            }
          },
          "404": {
            "description": "Поставщик не найден"
          }
        },
        "tags": [
          "suppliers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "SuppliersController_remove",
        "summary": "Удалить поставщика",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Поставщик успешно удален"
          },
          "404": {
            "description": "Поставщик не найден"
          }
        },
        "tags": [
          "suppliers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "ReceiptMaster API",
    "description": "Multi-tenant SaaS API for managing products, recipients, orders, receipts, and analytics.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "purchase_price_cents": {
            "type": "number"
          },
          "sale_price_cents": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "enum": [
              "UAH"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "purchase_price_cents",
          "sale_price_cents",
          "quantity",
          "currency",
          "created_at",
          "updated_at"
        ]
      },
...
      "CreateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название поставщика",
            "example": "ООО Поставщик"
          },
          "email": {
            "type": "string",
            "description": "Email поставщика",
            "example": "supplier@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Телефон поставщика",
            "example": "+380 50 123 4567"
          },
          "address": {
            "type": "string",
            "description": "Адрес поставщика",
            "example": "г. Киев, ул. Центральная, 1"
          },
          "contactPerson": {
            "type": "string",
            "description": "Контактное лицо",
            "example": "Иван Иванов"
          },
          "notes": {
            "type": "string",
            "description": "Заметки",
            "example": "Основной поставщик"
          },
          "productIds": {
            "description": "ID продуктов для привязки",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "Supplier": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contact_person": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "products": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "products",
          "created_at",
          "updated_at"
        ]
      },
      "PaginatedSupplierResponse": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Supplier"
            }
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "data"
        ]
      },
      "UpdateSupplierDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название поставщика",
            "example": "ООО Поставщик"
          },
          "email": {
            "type": "string",
            "description": "Email поставщика",
            "example": "supplier@example.com"
          },
          "phone": {
            "type": "string",
            "description": "Телефон поставщика",
            "example": "+380 50 123 4567"
          },
          "address": {
            "type": "string",
            "description": "Адрес поставщика",
            "example": "г. Киев, ул. Центральная, 1"
          },
          "contactPerson": {
            "type": "string",
            "description": "Контактное лицо",
            "example": "Иван Иванов"
          },
          "notes": {
            "type": "string",
            "description": "Заметки",
            "example": "Основной поставщик"
          },
          "productIds": {
            "description": "ID продуктов для привязки",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}